Fix inverted logic for --no-shuffle
This revision fixes an apparent inverted logic bug in processing of the "--no-shuffle" option.

The [text](https://freecontent.manning.com/wp-content/uploads/functional-reactive-programming-in-search-of-the-mythical-von-neumann-machine.pdf) describes the `linked-list` program timing with/without the option as:

```
time ./linked-list --no-shuffle
user 0m3.390s

time ./linked-list
user 1m19.563s
```

However we saw the behavior as inverse, e.g.:

```
$ time ./linked-list --no-shuffle

real	1m24.075s
user	1m24.054s
sys	0m0.020s

$ time ./linked-list

real	0m3.724s
user	0m3.708s
sys	0m0.017s
```

Prior to this fix, including the "--no-shuffle" option would make the program shuffle list entries,
which was the opposite of the intended effect.

This revision also adds an include for the string header as needed for `strcmp` to work correctly in GCC.
This was tested with GCC 7.5.0 on Ubuntu 18.04.6 LTS.